refactor(losses): move shared losses into losses.py
